shuffle and add error helpers
