+++ c/sedona-s2geography/src/lib.rs
@@ -15,5 +15,6 @@
 // specific language governing permissions and limitations
 // under the License.
 pub mod register;
-pub mod s2geography;
+pub mod utils;
 mod s2geography_c_bindgen;
+pub mod kernels;
