+++ rust/sedona/src/context.rs
@@ -253,7 +253,7 @@ impl SedonaContext {
         self.register_scalar_kernels(sedona_s2geography::register::scalar_kernels()?.into_iter())?;
 
         let sd_order_kernel = sd_order_lnglat::OrderLngLat::new(
-            sedona_s2geography::s2geography::s2_cell_id_from_lnglat,
+            sedona_s2geography::utils::s2_cell_id_from_lnglat,
         );
         self.register_scalar_kernels([("sd_order", sd_order_kernel)].into_iter())?;
